+++ src/test/ExSln2/c99/gen/issue58_VtableReturnsObjPointerEx.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "issue58_Bike.h"
+#include "issue58_IBikeProvider.h"